Change back current working directory directly after execution.
If the current working directory (which most likely is the directory containing the root composer.json and vendor folder of the project using PHPJasper) is not changed back after execution, this will break other code of the project depending on the current working directory being the project root.

See #291 for more information.